+++ CHANGELOG.md
@@ -6,7 +6,16 @@ Overall the core Framework files and API are expected to remain stable however t
 
 ## Next Release for NPM (Changes on Master Branch)
 
-* Fix `js/web-components/markdown-content.js` so that it shows it shows the correct background color and full CSS styles for Code Syntax when using [highlight.js](https://github.com/highlightjs/highlight.js). With previous releases this worked with the Framework version `js/controls/markdown-content.js`
+* Add React Component `<MarkdownContent>`
+  * Based on DataFormsJS Web Component `<markdown-content>`
+  * The DataFormsJS `<MarkdownContent>` React Component makes it easy to display markdown content from a web service or URL:
+  ~~~jsx
+  const url = "https://raw.githubusercontent.com/dataformsjs/dataformsjs/master/README.md"
+  <MarkdownContent url={url} />
+  <MarkdownContent url={url} isLoading={<IsLoading />} className="markdown" />
+  ~~~
+* Fix `js/web-components/markdown-content.js` so that it shows it shows the correct background color and full CSS styles for Code Syntax when using `highlight.js`. With previous releases this worked with the Framework version `js/controls/markdown-content.js`
+* Fix for both Framework and Web Component `<markdown-content>` where there were running `highlight.js` against the document rather than only elements under `<markdown-content>`.
 
 ## 5.1.1 (November 21, 2020)